requirements.txt
django
djangorestframework
djangorestframework-simplejwt
corsheaders
whitenoise
dotenv